Resolve conflicts in Form, pom, entities, application.properties and SQL schema; keep unused assets under NÃO-UTILIZADO.

+++ back/src/main/java/com/ayvy/api_java/dto/ProdutoImagemRequest.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 public class ProdutoImagemRequest {
 
     private String caminho;
-    private Integer ordem;
+    private Short ordem;
 }
